Various AKTable fixes, basically acting as a proxy for the Csound table instead of managing its own buffer. Use NSUInteger instead of int for array sizes, for consistency with Cocoa.

+++ AudioKit/Tables/Generators/AKExponentialTableGenerator.m
@@ -40,7 +40,7 @@
     [self addValue:value atIndex:index];
 }
 
-- (NSArray *)parametersWithSize:(int)size
+- (NSArray *)parametersWithSize:(NSUInteger)size
 {
     [points sortUsingComparator:^NSComparisonResult(NSArray *obj1, NSArray *obj2) {
         return [obj1[0] compare: obj2[0]];
